requirements.txt

Click>=6.0
python_bugzilla==1.2.2
six==1.10.0
